requirements.txt
yunhu
flask
python-dotenv
langdetect